LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/perspective-expected.txt


PASS perspective responsive to style changes
FAIL perspective clamped to 0px assert_equals: expected "0px" but got "100px"
FAIL perspective responsive to inherited changes assert_equals: expected "80px" but got "none"
